Move data-display outside organization-dashboard
