+++ client/app/components/data-display/daily-hours-chart/daily-hours-chart.component.css
